Add ts_stat_statements callbacks
Currently we finish the execution of some process utility statements
and don't execute other hooks in the chain.

Because that reason neither `ts_stat_statements` and
`pg_stat_statements` are able to track some utility statements, for
example COPY ... FROM.

To be able to track it on `ts_stat_statements` we're introducing some
callbacks in order to hook `pgss_store` from TimescaleDB and store
information about the execution of those statements.

In this PR we're also adding a new GUC `enable_tss_callbacks=fase` to
enable or disable the ability to hook `ts_stat_statements` from
TimescaleDB.